+++ scripts/setup-git-hooks.sh
@@ -22,7 +22,7 @@ fi
 echo "📦 Instalando hooks de pre-commit..."
 pre-commit install
 pre-commit install --hook-type prepare-commit-msg
-pre-commit install --hook-type commit-msg  
+pre-commit install --hook-type commit-msg
 pre-commit install --hook-type pre-push
 
 echo ""